x86/microcode: Sanitize __wait_for_cpus()
The code is too complicated for no reason.

The return value is pointless as this is a strict boolean.

It's way simpler to count down from num_online_cpus() and check for
zero.

The timeout argument is pointless as this is always one second.

Touching the NMI watchdog every 100ns does not make any sense, neither does
checking every 100ns. This is really not a hotpath operation.

Preload the atomic counter with the number of online CPUs and simplify the
whole timeout logic. Delay for one microsecond and touch the NMI watchdog
once per millisecond.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>